+++ packages/frontend/src/components/dialogs/ChatAuditLogDialog.tsx
@@ -39,8 +39,6 @@ function buildContextMenu(
   closeDialogCallback: DialogProps['onClose']
 ) {
   const tx = window.static_translate // don't use the i18n context here for now as this component is inefficient (rendered one menu for every message)
-  const isInfoOrCallInvitation =
-    message.isInfo || message.viewType === 'VideochatInvitation'
   return [
     // Show in Chat
     {
@@ -82,7 +80,7 @@ function buildContextMenu(
     // Reply
     // TODO also check `chat.canSend`, as with `showReply` in `Message.tsx`,
     // and double-check other items.
-    !isInfoOrCallInvitation && {
+    !message.isInfo && {
       label: tx('reply_noun'),
       action: () => {
         setQuoteInDraft(message.id)
@@ -91,7 +89,7 @@ function buildContextMenu(
     },
     // Reply privately -> only show in groups, don't show on info messages or outgoing messages
     isGroup &&
-      !isInfoOrCallInvitation &&
+      !message.isInfo &&
       message.fromId > C.DC_CONTACT_ID_LAST_SPECIAL && {
         label: tx('reply_privately'),
         action: () => {